+++ packages/pandas-gbq/tests/unit/test_gbq.py
@@ -47,6 +47,27 @@ def mock_get_credentials(*args, **kwargs):
     return mock_credentials, "default-project"
 
 
+@pytest.fixture
+def mock_service_account_credentials():
+    import google.oauth2.service_account
+
+    mock_credentials = mock.create_autospec(
+        google.oauth2.service_account.Credentials
+    )
+    return mock_credentials
+
+
+@pytest.fixture
+def mock_compute_engine_credentials():
+    import google.auth.compute_engine
+
+    mock_credentials = mock.create_autospec(
+        google.auth.compute_engine.Credentials
+    )
+    return mock_credentials
+
+
+@pytest.fixture
 def mock_get_user_credentials(*args, **kwargs):
     import google.auth.credentials
 
@@ -258,6 +279,36 @@ def test_read_gbq_with_no_project_id_given_should_fail(monkeypatch):
 def test_read_gbq_with_inferred_project_id(monkeypatch):
     df = gbq.read_gbq("SELECT 1", dialect="standard")
     assert df is not None
+
+
+def test_read_gbq_with_inferred_project_id_from_service_account_credentials(
+    mock_bigquery_client, mock_service_account_credentials
+):
+    mock_service_account_credentials.project_id = "service_account_project_id"
+    df = gbq.read_gbq(
+        "SELECT 1",
+        dialect="standard",
+        credentials=mock_service_account_credentials,
+    )
+    assert df is not None
+    mock_bigquery_client.query.assert_called_once_with(
+        "SELECT 1",
+        job_config=mock.ANY,
+        location=None,
+        project="service_account_project_id",
+    )
+
+
+def test_read_gbq_without_inferred_project_id_from_compute_engine_credentials(
+    mock_compute_engine_credentials
+):
+    with pytest.raises(ValueError) as exception:
+        gbq.read_gbq(
+            "SELECT 1",
+            dialect="standard",
+            credentials=mock_compute_engine_credentials,
+        )
+    assert "Could not determine project ID" in str(exception)
 
 
 def test_read_gbq_with_invalid_private_key_json_should_fail():