+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -13,6 +13,7 @@ Dependency updates
 
 - Update the minimum version of ``google-cloud-bigquery`` to 1.9.0.
   (:issue:`247`)
+- Update the minimum version of ``pandas`` to 0.19.0. (:issue:`262`)
 
 Internal changes
 ~~~~~~~~~~~~~~~~
@@ -23,7 +24,7 @@ Internal changes
 Enhancements
 ~~~~~~~~~~~~
 - Allow ``table_schema`` in :func:`to_gbq` to contain only a subset of columns,
-  with the rest being populated using the DataFrame dtypes (:issue:`218`) 
+  with the rest being populated using the DataFrame dtypes (:issue:`218`)
   (contributed by @johnpaton)
 - Read ``project_id`` in :func:`to_gbq` from provided ``credentials`` if
   available (contributed by @daureg)