+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -6,7 +6,11 @@ Changelog
 0.10.0 / TBD
 ------------
 
-- This fixes a bug where pandas-gbq could not upload an empty database. (:issue:`237`)
+- Fix a bug where pandas-gbq could not upload an empty DataFrame. (:issue:`237`)
+- Update the authentication credentials. **Note:** You may need to set
+  ``reauth=True`` in order to update your credentials to the most recent
+  version. This is required to use new functionality such as the BigQuery
+  Storage API. (:issue:`267`)
 
 Dependency updates
 ~~~~~~~~~~~~~~~~~~